Create README.md
This directory contains a copy of an example parameter file to run the Extended Cardiac Torso (XCAT) Phantom image software.  

The parameter (.par) file is the input file to the phantom was used to create the XCAT phantom images (mode 0) and the ASCII files which describe position vectors frame to frame (mode 4).  

The values in the .par file were chosen to emulate an emission modality, such as PET or SPECT imaging (gamma photons have an energy of 140keV, the energy produced by gamma photons using Technetium-99m, a metastable nuclear isomer of technetium-99, 99mTc) .  

There is no breath motion simulated.